refactor game app
