+++ wordle_api/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from wordle_api.routers import game
+from wordle_api.game import router
 from wordle_api.user.router import router as user_router
 # from wordle_api.auth.asdf import router as auth_router
 
@@ -17,5 +17,5 @@ app.add_middleware(
     allow_headers=["*"],
 )
 app.include_router(user_router)
-app.include_router(game.router)
+app.include_router(router.router)
 # app.include_router(auth_router)
